Restructure tests/compiler/ to mirror deplodock/compiler/
Top-level tests/compiler/ had grown to 43 files mixing IR datatypes,
pass tests, backend tests, trace tests, CLI tests, and end-to-end
suites. Reorganise into subdirs that mirror the source layout, merge
near-duplicates, and promote the most-copied helper to a shared
fixture.

Subdirs (mirroring deplodock/compiler/):

- ir/ — Graph/Tensor/Node, splice, structural_key, Hints, IndexMap,
  LoopOp SSA body (was test_kernel_op.py), shape inference, provenance,
  Dim/dynamic shapes, real-trace fixture
- passes/ — joins the existing rule-test layout with the moved
  test_matcher, test_pipeline_semantics (was test_lower), tile naming,
  knob pinning, blocked GEMM, and the new test_lowering_accuracy.py
  (collapses async_copy + double_buffer + pad_smem + tma — same shape,
  one test per pass)
- pipeline/ — rule_diff + merged test_dump.py (was test_dump_dot +
  test_provenance_dump); search/ gets test_tune_accuracy
- backend/ (new) — dtype_cuda/numpy, emit, loader, nvcc_compile,
  program, torch_ref, bench_worker_recovery
- trace/ (new) — torch tracer (test_torch_trace_ops.py renamed,
  absorbed the 2 smoke cases that were in test_torch_trace.py)
- cli/ (new) — compile, knobs, run (CLI suffix redundant in cli/)
- e2e/ (new) — accuracy, block, attention_chains, pipeline,
  ops_vs_torch (was misleadingly named test_torch_ops.py)

Merges:

- test_shape_inference_symbolic.py → ir/test_shape_inference.py
- test_provenance_splice.py → ir/test_provenance.py
- test_provenance_dump.py → pipeline/test_dump.py
- test_async_copy / double_buffer / pad_smem / tma_swizzle accuracy
  files → passes/test_lowering_accuracy.py (saves ~330 LOC of
  duplicated matmul-graph + assert-close scaffolding)

Deletions:

- test_rewriter.py — silu decomp smoke fully covered by
  passes/test_decompose_rules.py
- test_torch_trace.py — 2 smoke tests already supplemented by
  test_torch_trace_ops.py per the latter's docstring

Shared fixture:

- tests/compiler/conftest.py grows matmul_graph(m,k,n) — the
  (m,k)@(k,n)→(m,n) builder that was copy-pasted across 8 callers.
  Migrated test_pipeline / test_emit / test_lowering_tma /
  test_lowering_accuracy.

Style cleanup:

- Replaced 3 local pytest.mark.skipif(not torch.cuda.is_available(),...)
  definitions with `from ..conftest import requires_cuda` for
  consistency with the rest of the suite
- Fixed test_bench_worker_recovery.py's absolute
  `from tests.compiler.conftest import ...` to relative

Docs:

- tests/ARCHITECTURE.md rewritten to describe the new compiler subdirs
- tests/compiler/passes/ARCHITECTURE.md updated file list + paths

867 compiler tests collect (down from 872 — 5 deletions), all 1323
tests pass, make lint clean.
