+++ tests/compiler/backend/test_program.py
@@ -5,7 +5,7 @@ from deplodock.compiler.graph import Graph, Tensor
 from deplodock.compiler.ir.base import InputOp
 from deplodock.compiler.ir.cuda import CudaOp
 
-from .conftest import requires_cuda
+from ..conftest import requires_cuda
 
 EW_ADD_SOURCE = """
 extern "C" __global__ void ew_add(const float* A, const float* B, float* C) {
