+++ tests/compiler/cli/test_compile.py
