+++ tests/compiler/ir/test_graph_splice.py
