+++ tests/compiler/ir/test_graph_structural_key.py
