+++ tests/compiler/ir/test_hints.py
