+++ tests/compiler/ir/test_loop_op.py
