+++ tests/compiler/pipeline/search/test_tune_accuracy.py
@@ -34,7 +34,7 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-from .conftest import requires_cuda
+from ...conftest import requires_cuda
 
 # Small-shape variants of the perf suite. Dimensions are picked so each
 # planner branch (matmul, cooperative-K, multi-Accum matmul, multi-stmt
