+++ tests/compiler/pipeline/test_rule_diff.py
