--- tests/compiler/test_shape_inference.py
@@ -1,82 +1,0 @@
-"""Unit tests for Op.infer_output_shape."""
-
-from deplodock.compiler.ir.frontend.ir import (
-    CatOp,
-    LinearOp,
-    MatmulOp,
-    MeanOp,
-    ReshapeOp,
-    SdpaOp,
-    SliceOp,
-    TransposeOp,
-    UnsqueezeOp,
-)
-from deplodock.compiler.ir.tensor.ir import ElementwiseOp, ReduceOp
-
-
-def test_elementwise_infer_matching_shapes():
-    """ElementwiseOp is rank-preserving: matching inputs pass through."""
-    op = ElementwiseOp("multiply")
-    assert op.infer_output_shape([(1, 28, 8, 128), (1, 28, 8, 128)]) == (1, 28, 8, 128)
-
-
-def test_elementwise_infer_rejects_mismatched_shapes():
-    """ElementwiseOp refuses to broadcast — callers must wrap in IndexMapOp first."""
-    op = ElementwiseOp("multiply")
-    try:
-        op.infer_output_shape([(1, 28, 8, 128), (1, 1, 8, 128)])
-    except ValueError as e:
-        assert "must all match" in str(e)
-    else:
-        raise AssertionError("should have raised on mismatched shapes")
-
-
-def test_reduce_keepdim():
-    assert ReduceOp("sum", axis=-1).infer_output_shape([(1, 8, 128)]) == (1, 8, 1)
-    assert ReduceOp("maximum", axis=1).infer_output_shape([(4, 8, 128)]) == (4, 1, 128)
-
-
-def test_mean_keepdim():
-    assert MeanOp(axis=-1).infer_output_shape([(1, 8, 3584)]) == (1, 8, 1)
-
-
-def test_transpose_swap():
-    assert TransposeOp(axes=(1, 2)).infer_output_shape([(1, 8, 28, 128)]) == (1, 28, 8, 128)
-
-
-def test_transpose_full_permute():
-    assert TransposeOp(axes=(0, 2, 1, 3)).infer_output_shape([(1, 8, 28, 128)]) == (1, 28, 8, 128)
-
-
-def test_reshape_returns_op_shape():
-    assert ReshapeOp(shape=(1, 8, 3584)).infer_output_shape([(1, 28, 8, 128)]) == (1, 8, 3584)
-
-
-def test_unsqueeze_inserts_dim():
-    assert UnsqueezeOp(dim=1).infer_output_shape([(1, 8, 128)]) == (1, 1, 8, 128)
-    assert UnsqueezeOp(dim=-1).infer_output_shape([(1, 8, 128)]) == (1, 8, 128, 1)
-
-
-def test_slice_returns_op_shape():
-    assert SliceOp(shape=(1, 28, 8, 64)).infer_output_shape([(1, 28, 8, 128)]) == (1, 28, 8, 64)
-
-
-def test_cat_sums_last_dim():
-    assert CatOp().infer_output_shape([(1, 28, 8, 64), (1, 28, 8, 64)]) == (1, 28, 8, 128)
-
-
-def test_linear_replaces_last_with_out_features():
-    assert LinearOp().infer_output_shape([(1, 8, 3584), (3584, 3584)]) == (1, 8, 3584)
-    # K projection: out_features=512 (the -2 dim of weight)
-    assert LinearOp().infer_output_shape([(1, 8, 3584), (512, 3584)]) == (1, 8, 512)
-
-
-def test_matmul_standard():
-    assert MatmulOp().infer_output_shape([(8, 16), (16, 32)]) == (8, 32)
-
-
-def test_sdpa_uses_v_last_dim():
-    q = (1, 28, 8, 64)
-    k = (1, 4, 8, 64)
-    v = (1, 4, 8, 64)
-    assert SdpaOp().infer_output_shape([q, k, v]) == (1, 28, 8, 64)
